% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chinook.R
\docType{data}
\name{chinook}
\alias{chinook}
\title{Chinook Salmon Escapement estimates for California's Central Valley

A simplified version of the Chinook Salmon Escapement estimates for California's Central Valley for winter, spring, fall, and late-fall run.
Total escapement includes hatchery and in-river fish.}
\format{
An object of class \code{"data.frame"}
\describe{
\item{year}{A value between 1967 and 2019}
\item{Run}{Run type}
\item{Escapement}{Total escapement estimate}
}
\usage{
data(chinook)
}
\description{
Chinook Salmon Escapement estimates for California's Central Valley

A simplified version of the Chinook Salmon Escapement estimates for California's Central Valley for winter, spring, fall, and late-fall run.
Total escapement includes hatchery and in-river fish.
}
\examples{
data(chinook)
head(chinook)

}
\references{
California Department of Fish and Wildlife, California Central Valley Chinook Population Database Report,GrandTab 2020.05.22
}
\keyword{datasets}
