% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfecol.R
\name{get_pal}
\alias{get_pal}
\title{Select SFE color palette}
\usage{
get_pal(name)
}
\arguments{
\item{name}{Name of desired palette. Choices are: \code{smelt}, \code{chinook}, and \code{southbay}}
}
\value{
A vector of colours.
}
\description{
This function returns color palettes derived from photos of the San Francisco Estuary and its wildlife
}
\examples{
get_pal("smelt")
get_pal("chinook")
get_pal("southbay")
}
