% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfecol.R
\name{sfe_pal}
\alias{sfe_pal}
\title{Select SFE color palette}
\usage{
sfe_pal(name)
}
\arguments{
\item{name}{Name of desired palette. Choices are: \code{delsme}, \code{chinook},  \code{chinookcarc}, and \code{southbay}}
}
\value{
A vector of colors.
}
\description{
This function returns color palettes derived from photos of the San Francisco Estuary and its wildlife
}
\examples{
get_pal("delsme")
get_pal("chinook")
get_pal("chinookcarc")
get_pal("southbay")

}
