% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfecol.R
\docType{data}
\name{sfe_palettes}
\alias{sfe_palettes}
\title{sfecol}
\format{
An object of class \code{list} of length 3.
}
\usage{
sfe_palettes
}
\description{
Color palettes based on the San Francisco Estuary
}
\examples{
data()

}
\keyword{datasets}
