% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfecol.R
\name{sfe_pal}
\alias{sfe_pal}
\title{Select SFE color palette}
\usage{
sfe_pal(name)
}
\arguments{
\item{name}{Name of desired palette. Choices are: \code{delsme}, \code{chinook},  \code{chinookcarc}, \code{southbay},
\code{baslu},\code{petaluma}, \code{napa},\code{elpastel}}
}
\value{
A vector of colors.
}
\description{
This function returns color palettes derived from photos of the San Francisco Estuary.
}
\examples{
sfe_pal("delsme")
sfe_pal("chinook")
sfe_pal("chinookcarc")
sfe_pal("southbay")
sfe_pal("petaluma")
sfe_pal("napa")
sfe_pal("baslu")
sfe_pal("elpastel")

}
