% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfecol.R
\docType{data}
\name{sfe_palettes}
\alias{sfe_palettes}
\title{sfecol}
\format{
An object of class \code{list} of length 19.
}
\usage{
sfe_palettes
}
\description{
Color palettes based on the San Francisco Estuary
Complete list of palettes

These are colour palettes derived from photos of birds native to New Zealand. Use \code{\link{get_pal}} extract palettes.
}
\keyword{datasets}
