% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sfecol.R
\docType{data}
\name{sfe_palettes}
\alias{sfe_palettes}
\title{Color palettes based on the San Francisco Estuary}
\format{
An object of class \code{list} of length 5.
}
\usage{
sfe_palettes
}
\description{
This package provides color palettes based on images from the San Francisco Estuary. Use \code{\link{sfe_pal}} extract palettes.
Package based on the excellent manu package (https://g-thomson.github.io/Manu/) by Dr Tara McAllister and Karthik Ram
Complete list of palettes
}
\keyword{datasets}
